added parseComposeToProduction
