+++ src/releaseFiles/compose/utils.ts
@@ -1,0 +1,5 @@
+import { Manifest } from "../manifest/types";
+
+export function getIsCore(manifest: Manifest): boolean {
+  return manifest.type === "dncore";
+}
