+++ test/validation/compose/bad-docker-compose.yml
@@ -8,6 +8,7 @@ services:
         UPSTREAM_VERSION: v2.2.0
     volumes:
       - "beacon-data:/root/.lighthouse"
+      - "/host/path:/container/path"
     ports:
       - 9000/tcp
       - 9000/udp
