+++ test/validation/compose/good-docker-compose.yml
@@ -8,7 +8,6 @@ services:
         UPSTREAM_VERSION: v2.2.0
     volumes:
       - "beacon-data:/root/.lighthouse"
-      - "/host/path:/root/.lighthouse"
     ports:
       - 9000/tcp
       - 9000/udp
